Extracted FullStoreChangeStream from LabelScanProvider
