+++ community/kernel/src/main/java/org/neo4j/kernel/impl/api/scan/LabelScanStoreProvider.java
@@ -19,25 +19,15 @@
  */
 package org.neo4j.kernel.impl.api.scan;
 
-import org.apache.commons.lang3.ArrayUtils;
-
 import java.io.File;
 import java.io.IOException;
-import java.util.function.Supplier;
-
-import org.neo4j.helpers.collection.Visitor;
 import org.neo4j.kernel.api.labelscan.LabelScanStore;
 import org.neo4j.kernel.api.labelscan.LabelScanWriter;
-import org.neo4j.kernel.api.labelscan.NodeLabelUpdate;
 import org.neo4j.kernel.extension.KernelExtensionFactory;
 import org.neo4j.kernel.extension.KernelExtensions;
-import org.neo4j.kernel.impl.api.index.IndexStoreView;
-import org.neo4j.kernel.impl.api.index.StoreScan;
 import org.neo4j.kernel.lifecycle.LifeSupport;
 import org.neo4j.kernel.lifecycle.Lifecycle;
 import org.neo4j.kernel.lifecycle.LifecycleAdapter;
-
-import static org.neo4j.function.Predicates.ALWAYS_TRUE_INT;
 
 /**
  * Used by a {@link KernelExtensions} to provide access a {@link LabelScanStore} and prioritize against other.
@@ -86,56 +76,11 @@ public class LabelScanStoreProvider extends LifecycleAdapter implements Comparab
         return getClass().getSimpleName() + "[" + labelScanStore + ", prio:" + priority + "]";
     }
 
-    public interface FullStoreChangeStream
-    {
-        long applyTo( LabelScanWriter writer ) throws IOException;
-    }
-
-    public static final FullStoreChangeStream EMPTY = writer -> 0;
-
-    public static FullStoreChangeStream fullStoreLabelUpdateStream( Supplier<IndexStoreView> lazyIndexStoreView )
-    {
-        // IndexStoreView provided as supplier because we only actually have that dependency available
-        // when it's time to rebuilt, not when we construct this object
-        return new FullLabelStream( lazyIndexStoreView );
-    }
-
     public static long rebuild( LabelScanStore store, FullStoreChangeStream fullStoreStream ) throws IOException
     {
         try ( LabelScanWriter writer = store.newWriter() )
         {
             return fullStoreStream.applyTo( writer );
-        }
-    }
-
-    private static class FullLabelStream implements FullStoreChangeStream, Visitor<NodeLabelUpdate,IOException>
-    {
-        private final Supplier<IndexStoreView> lazyIndexStoreView;
-        private LabelScanWriter writer;
-        private long count;
-
-        public FullLabelStream( Supplier<IndexStoreView> lazyIndexStoreView )
-        {
-            this.lazyIndexStoreView = lazyIndexStoreView;
-        }
-
-        @Override
-        public long applyTo( LabelScanWriter writer ) throws IOException
-        {
-            // Keep the write for using it in visit
-            this.writer = writer;
-            IndexStoreView view = lazyIndexStoreView.get();
-            StoreScan<IOException> scan = view.visitNodes( ArrayUtils.EMPTY_INT_ARRAY, ALWAYS_TRUE_INT, null, this );
-            scan.run();
-            return count;
-        }
-
-        @Override
-        public boolean visit( NodeLabelUpdate update ) throws IOException
-        {
-            writer.write( update );
-            count++;
-            return false;
         }
     }
 }
